cart and orders

+++ Edutor/models/Product.js
@@ -6,7 +6,7 @@ const sequelize = require('sequelize');
 const Product = db.define('product',
     {
         name: { type: Sequelize.STRING },
-        price: { type: Sequelize.DECIMAL(10,2) }
+        price: { type: Sequelize.INTEGER}
     });
